Game and GameComponents
Added a Game Core Class which is supposed to be used as the core loop.
It uses Components which it updates automatically. Components can be
derived from the GameComponent base class or one of the premade ones.
Added EntityManager and LayeredEntityManager as examples.
Switched Hymns code to use the new structure.
LayeredEntityManager not on use yet -> Messed up draw order.