Fixed ranged Deposit Claim Bug
Was due to the pathfinding system not finding a path to the closest
deposit.
Why that is I don't know... maybe just to far away.
But the Behaviors relied on the entity just doing what it should.
In the end the entity thought it was at the target location.
So behaviors should always assert their own correctness in the future.