Fixed Minion Master jumping to enemy position
Unit:moveTo and Unit:moveToTarget were changed to have targetPosition refer directly to the vector representing the position of another entity. Then, when Entity:setPosition was called, the target position was getting modified, in effect changing the position of the target entity.

Solved by not sharing references to the same vector instance.